Added patches from official OSX installer
The official GTK+ OSX installer is a python app that downloads and
applies a set of patches from gnome.org. I've included the patches
specific to 2.24.9, the latest 2.x version it supports.

These patches mostly concern Quartz integration and the new version of
Cups in Mac OS X 10.8 (Mountain Lion).